Add fct_replace function
Add convenience functions `fct_replace` and `fct_replace_all` which replace factor levels using regular expressions.

This adds the `stringr` dependency. So it may be better to use `sub` and `gsub` instead.

I'm also not sure if this function is worth it. It's a common use case (at least for me), but is a just a thin wrapper around `fct_relabel`.